More Optimization and Annoyance Improvements
Redid the underground section to have less non-full blocks, changed glass panes in above ground blocks to be a special block specific to that color. If more optimization is needed, remove some blocks from the underground

Some annoyances updated to actually work, although most are still untested and are probably broken. This version has everything working to the best of my knowledge