Quick fixes to yesterday's major thingy
I probably should have just waited for this all to be done before committing. Idk I was just so happy to finally fix this dumb lag issue after like 6 days of doing nothing but working on it.

Oh wait that's right I'm not supposed to put actual stuff here. uhhh.... did the thing